// stashmate/src/types/schema.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          extensions?: Json
          operationName?: string
          query?: string
          variables?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      collections: {
        Row: {
          acquired_date: string
          category: string
          cond: string | null
          cost: number
          id: number
          name: string
          owner_id: string | null
          profit: number
          qty: number | null
          source: string | null
          status: number
          value: number
        }
        Insert: {
          acquired_date: string
          category: string
          cond?: string | null
          cost: number
          id?: number
          name: string
          owner_id?: string | null
          profit?: number
          qty?: number | null
          source?: string | null
          status: number
          value: number
        }
        Update: {
          acquired_date?: string
          category?: string
          cond?: string | null
          cost?: number
          id?: number
          name?: string
          owner_id?: string | null
          profit?: number
          qty?: number | null
          source?: string | null
          status?: number
          value?: number
        }
        Relationships: [
          {
            foreignKeyName: "collections_owner_id_fkey"
            columns: ["owner_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      items: {
        Row: {
          collection_id: number
          condition: string | null
          cost: number
          created_at: string
          id: number
          name: string | null
          price: number
          profit: number
          source: string
          status: number
        }
        Insert: {
          collection_id: number
          condition?: string | null
          cost: number
          created_at: string
          id?: number
          name?: string | null
          price: number
          profit: number
          source: string
          status: number
        }
        Update: {
          collection_id?: number
          condition?: string | null
          cost?: number
          created_at?: string
          id?: number
          name?: string | null
          price?: number
          profit?: number
          source?: string
          status?: number
        }
        Relationships: [
          {
            foreignKeyName: "items_collection_id_fkey"
            columns: ["collection_id"]
            isOneToOne: false
            referencedRelation: "collections"
            referencedColumns: ["id"]
          },
        ]
      }
      users: {
        Row: {
          created_at: string
          email: string | null
          id: string
        }
        Insert: {
          created_at?: string
          email?: string | null
          id: string
        }
        Update: {
          created_at?: string
          email?: string | null
          id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never